@@ -1,114 +1,18 @@
 // Copyright 2021-2022, Offchain Labs, Inc.
 // For license information, see https://github.com/nitro/blob/master/LICENSE
 
+#[cfg(feature = "native")]
 use crate::kzg::ETHEREUM_KZG_SETTINGS;
 use arbutil::PreimageType;
+#[cfg(feature = "native")]
 use c_kzg::{Blob, KzgCommitment};
 use digest::Digest;
 use eyre::{eyre, Result};
 use serde::{Deserialize, Serialize};
 use sha2::Sha256;
 use sha3::Keccak256;
-use std::{
-    borrow::Borrow,
-    convert::TryInto,
-    fmt,
-    fs::File,
-    io::Read,
-    ops::{Deref, DerefMut},
-    path::Path,
-};
-use wasmparser::{TableType, Type};
-
-/// cbindgen:field-names=[bytes]
-#[derive(Default, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
-#[repr(C)]
-pub struct Bytes32(pub [u8; 32]);
-
-impl Deref for Bytes32 {
-    type Target = [u8; 32];
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl DerefMut for Bytes32 {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.0
-    }
-}
-
-impl AsRef<[u8]> for Bytes32 {
-    fn as_ref(&self) -> &[u8] {
-        &self.0
-    }
-}
-
-impl Borrow<[u8]> for Bytes32 {
-    fn borrow(&self) -> &[u8] {
-        &self.0
-    }
-}
-
-impl From<[u8; 32]> for Bytes32 {
-    fn from(x: [u8; 32]) -> Self {
-        Self(x)
-    }
-}
-
-impl From<u32> for Bytes32 {
-    fn from(x: u32) -> Self {
-        let mut b = [0u8; 32];
-        b[(32 - 4)..].copy_from_slice(&x.to_be_bytes());
-        Self(b)
-    }
-}
-
-impl From<u64> for Bytes32 {
-    fn from(x: u64) -> Self {
-        let mut b = [0u8; 32];
-        b[(32 - 8)..].copy_from_slice(&x.to_be_bytes());
-        Self(b)
-    }
-}
-
-impl From<usize> for Bytes32 {
-    fn from(x: usize) -> Self {
-        let mut b = [0u8; 32];
-        b[(32 - (usize::BITS as usize / 8))..].copy_from_slice(&x.to_be_bytes());
-        Self(b)
-    }
-}
-
-impl IntoIterator for Bytes32 {
-    type Item = u8;
-    type IntoIter = std::array::IntoIter<u8, 32>;
-
-    fn into_iter(self) -> Self::IntoIter {
-        IntoIterator::into_iter(self.0)
-    }
-}
-
-type GenericBytes32 = digest::generic_array::GenericArray<u8, digest::generic_array::typenum::U32>;
-
-impl From<GenericBytes32> for Bytes32 {
-    fn from(x: GenericBytes32) -> Self {
-        <[u8; 32]>::from(x).into()
-    }
-}
-
-impl fmt::Display for Bytes32 {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", hex::encode(self))
-    }
-}
-
-impl fmt::Debug for Bytes32 {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", hex::encode(self))
-    }
-}
+use std::{borrow::Borrow, convert::TryInto, fmt, fs::File, io::Read, ops::Deref, path::Path};
+use wasmparser::{RefType, TableType};
 
 /// A Vec<u8> allocated with libc::malloc
 pub struct CBytes {
@@ -171,23 +75,41 @@ impl From<&[u8]> for CBytes {
 unsafe impl Send for CBytes {}
 unsafe impl Sync for CBytes {}
 
+/// Unfortunately, [`wasmparser::RefType`] isn't serde and its contents aren't public.
+/// This type enables serde via a 1:1 transmute.
 #[derive(Serialize, Deserialize)]
-#[serde(remote = "Type")]
-enum RemoteType {
-    I32,
-    I64,
-    F32,
-    F64,
-    V128,
-    FuncRef,
-    ExternRef,
+struct RemoteRefType(pub [u8; 3]);
+
+impl From<RefType> for RemoteRefType {
+    fn from(value: RefType) -> Self {
+        unsafe { std::mem::transmute(value) }
+    }
+}
+
+impl From<RemoteRefType> for RefType {
+    fn from(value: RemoteRefType) -> Self {
+        unsafe { std::mem::transmute(value) }
+    }
+}
+
+mod remote_convert {
+    use super::{RefType, RemoteRefType};
+    use serde::{Deserialize, Deserializer, Serialize, Serializer};
+
+    pub fn serialize<S: Serializer>(value: &RefType, serializer: S) -> Result<S::Ok, S::Error> {
+        RemoteRefType::from(*value).serialize(serializer)
+    }
+
+    pub fn deserialize<'de, D: Deserializer<'de>>(deserializer: D) -> Result<RefType, D::Error> {
+        Ok(RemoteRefType::deserialize(deserializer)?.into())
+    }
 }
 
 #[derive(Serialize, Deserialize)]
 #[serde(remote = "TableType")]
 pub struct RemoteTableType {
-    #[serde(with = "RemoteType")]
-    pub element_type: Type,
+    #[serde(with = "remote_convert")]
+    pub element_type: RefType,
     pub initial: u32,
     pub maximum: Option<u32>,
 }
@@ -268,6 +190,7 @@ pub fn split_import(qualified: &str) -> Result<(&str, &str)> {
     Ok((module, name))
 }
 
+#[cfg(feature = "native")]
 pub fn hash_preimage(preimage: &[u8], ty: PreimageType) -> Result<[u8; 32]> {
     match ty {
         PreimageType::Keccak256 => Ok(Keccak256::digest(preimage).into()),
