@@ -10,6 +10,7 @@ import (
 	"github.com/ethereum/go-ethereum/core"
 	"github.com/ethereum/go-ethereum/core/vm"
 	"github.com/ethereum/go-ethereum/log"
+	"github.com/ethereum/go-ethereum/params"
 	"github.com/offchainlabs/nitro/arbos"
 	"github.com/offchainlabs/nitro/precompiles"
 )
@@ -55,14 +56,24 @@ func init() {
 		vm.PrecompiledContractsArbitrum[k] = v
 	}
 
+	for k, v := range vm.PrecompiledContractsCancun {
+		vm.PrecompiledAddressesArbOS30 = append(vm.PrecompiledAddressesArbOS30, k)
+		vm.PrecompiledContractsArbOS30[k] = v
+	}
+
 	precompileErrors := make(map[[4]byte]abi.Error)
 	for addr, precompile := range precompiles.Precompiles() {
 		for _, errABI := range precompile.Precompile().GetErrorABIs() {
 			precompileErrors[[4]byte(errABI.ID.Bytes())] = errABI
 		}
 		var wrapped vm.AdvancedPrecompile = ArbosPrecompileWrapper{precompile}
-		vm.PrecompiledContractsArbitrum[addr] = wrapped
-		vm.PrecompiledAddressesArbitrum = append(vm.PrecompiledAddressesArbitrum, addr)
+		vm.PrecompiledContractsArbOS30[addr] = wrapped
+		vm.PrecompiledAddressesArbOS30 = append(vm.PrecompiledAddressesArbOS30, addr)
+
+		if precompile.Precompile().ArbosVersion() < params.ArbosVersion_Stylus {
+			vm.PrecompiledContractsArbitrum[addr] = wrapped
+			vm.PrecompiledAddressesArbitrum = append(vm.PrecompiledAddressesArbitrum, addr)
+		}
 	}
 
 	for addr, precompile := range vm.PrecompiledContractsArbitrum {
