@@ -1,4 +1,4 @@
-// Copyright 2021-2022, Offchain Labs, Inc.
+// Copyright 2021-2024, Offchain Labs, Inc.
 // For license information, see https://github.com/nitro/blob/master/LICENSE
 
 package precompiles
@@ -11,6 +11,9 @@ import (
 	"math/big"
 
 	"github.com/offchainlabs/nitro/arbos/l1pricing"
+	"github.com/offchainlabs/nitro/arbos/programs"
+	"github.com/offchainlabs/nitro/util/arbmath"
+	am "github.com/offchainlabs/nitro/util/arbmath"
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/params"
@@ -164,6 +167,129 @@ func (con ArbOwner) ReleaseL1PricerSurplusFunds(c ctx, evm mech, maxWeiToRelease
 		return nil, err
 	}
 	return weiToTransfer, nil
+}
+
+// Sets the amount of ink 1 gas buys
+func (con ArbOwner) SetInkPrice(c ctx, evm mech, inkPrice uint32) error {
+	params, err := c.State.Programs().Params()
+	if err != nil {
+		return err
+	}
+	ink, err := arbmath.IntToUint24(inkPrice)
+	if err != nil || ink == 0 {
+		return errors.New("ink price must be a positive uint24")
+	}
+	params.InkPrice = ink
+	return params.Save()
+}
+
+// Sets the maximum depth (in wasm words) a wasm stack may grow
+func (con ArbOwner) SetWasmMaxStackDepth(c ctx, evm mech, depth uint32) error {
+	params, err := c.State.Programs().Params()
+	if err != nil {
+		return err
+	}
+	params.MaxStackDepth = depth
+	return params.Save()
+}
+
+// Gets the number of free wasm pages a tx gets
+func (con ArbOwner) SetWasmFreePages(c ctx, evm mech, pages uint16) error {
+	params, err := c.State.Programs().Params()
+	if err != nil {
+		return err
+	}
+	params.FreePages = pages
+	return params.Save()
+}
+
+// Sets the base cost of each additional wasm page
+func (con ArbOwner) SetWasmPageGas(c ctx, evm mech, gas uint16) error {
+	params, err := c.State.Programs().Params()
+	if err != nil {
+		return err
+	}
+	params.PageGas = gas
+	return params.Save()
+}
+
+// Sets the initial number of pages a wasm may allocate
+func (con ArbOwner) SetWasmPageLimit(c ctx, evm mech, limit uint16) error {
+	params, err := c.State.Programs().Params()
+	if err != nil {
+		return err
+	}
+	params.PageLimit = limit
+	return params.Save()
+}
+
+// Sets the minimum costs to invoke a program
+func (con ArbOwner) SetWasmMinInitGas(c ctx, _ mech, gas, cached uint64) error {
+	params, err := c.State.Programs().Params()
+	if err != nil {
+		return err
+	}
+	params.MinInitGas = am.SaturatingUUCast[uint8](am.DivCeil(gas, programs.MinInitGasUnits))
+	params.MinCachedInitGas = am.SaturatingUUCast[uint8](am.DivCeil(cached, programs.MinCachedGasUnits))
+	return params.Save()
+}
+
+// Sets the linear adjustment made to program init costs
+func (con ArbOwner) SetWasmInitCostScalar(c ctx, _ mech, percent uint64) error {
+	params, err := c.State.Programs().Params()
+	if err != nil {
+		return err
+	}
+	params.InitCostScalar = am.SaturatingUUCast[uint8](am.DivCeil(percent, programs.CostScalarPercent))
+	return params.Save()
+}
+
+// Sets the number of days after which programs deactivate
+func (con ArbOwner) SetWasmExpiryDays(c ctx, _ mech, days uint16) error {
+	params, err := c.State.Programs().Params()
+	if err != nil {
+		return err
+	}
+	params.ExpiryDays = days
+	return params.Save()
+}
+
+// Sets the age a program must be to perform a keepalive
+func (con ArbOwner) SetWasmKeepaliveDays(c ctx, _ mech, days uint16) error {
+	params, err := c.State.Programs().Params()
+	if err != nil {
+		return err
+	}
+	params.KeepaliveDays = days
+	return params.Save()
+}
+
+// Sets the number of extra programs ArbOS caches during a given block
+func (con ArbOwner) SetWasmBlockCacheSize(c ctx, _ mech, count uint16) error {
+	params, err := c.State.Programs().Params()
+	if err != nil {
+		return err
+	}
+	params.BlockCacheSize = count
+	return params.Save()
+}
+
+// Adds account as a wasm cache manager
+func (con ArbOwner) AddWasmCacheManager(c ctx, _ mech, manager addr) error {
+	return c.State.Programs().CacheManagers().Add(manager)
+}
+
+// Removes account from the list of wasm cache managers
+func (con ArbOwner) RemoveWasmCacheManager(c ctx, _ mech, manager addr) error {
+	managers := c.State.Programs().CacheManagers()
+	isMember, err := managers.IsMember(manager)
+	if err != nil {
+		return err
+	}
+	if !isMember {
+		return errors.New("tried to remove non-manager")
+	}
+	return managers.Remove(manager, c.State.ArbOSVersion())
 }
 
 func (con ArbOwner) SetChainConfig(c ctx, evm mech, serializedChainConfig []byte) error {
