@@ -193,13 +193,15 @@ func TestPrecompilesPerArbosVersion(t *testing.T) {
 		10: 2,
 		11: 4,
 		20: 8,
+		30: 38,
 	}
 
 	precompiles := Precompiles()
 	newMethodsPerArbosVersion := make(map[uint64]int)
 	for _, precompile := range precompiles {
 		for _, method := range precompile.Precompile().methods {
-			newMethodsPerArbosVersion[method.arbosVersion]++
+			version := arbmath.MaxInt(method.arbosVersion, precompile.Precompile().arbosVersion)
+			newMethodsPerArbosVersion[version]++
 		}
 	}
 
