@@ -1,18 +1,12 @@
-// Copyright 2021-2022, Offchain Labs, Inc.
-// For license information, see https://github.com/nitro/blob/master/LICENSE
+// Copyright 2021-2023, Offchain Labs, Inc.
+// For license information, see https://github.com/OffchainLabs/nitro/blob/master/LICENSE
 
 //go:build challengetest
 // +build challengetest
 
-//
-// Copyright 2021-2022, Offchain Labs, Inc. All rights reserved.
-//
-
 package arbtest
 
-import (
-	"testing"
-)
+import "testing"
 
 func TestChallengeManagerFullAsserterIncorrect(t *testing.T) {
 	t.Parallel()
