@@ -68,7 +68,7 @@ func TestCustomSolidityErrors(t *testing.T) {
 		Fatal(t, "customRevert call should have errored")
 	}
 	observedMessage := customError.Error()
-	expectedError := "Custom(1024, This spider family wards off bugs: /\\oo/\\ //\\(oo)/\\ /\\oo/\\, true)"
+	expectedError := "Custom(1024, This spider family wards off bugs: /\\oo/\\ //\\(oo)//\\ /\\oo/\\, true)"
 	// The first error is server side. The second error is client side ABI decoding.
 	expectedMessage := fmt.Sprintf("execution reverted: error %v: %v", expectedError, expectedError)
 	if observedMessage != expectedMessage {
