@@ -1,4 +1,4 @@
-// Copyright 2021-2022, Offchain Labs, Inc.
+// Copyright 2021-2024, Offchain Labs, Inc.
 // For license information, see https://github.com/nitro/blob/master/LICENSE
 
 package arbmath
@@ -6,8 +6,10 @@ package arbmath
 import "math/big"
 
 type Bips int64
+type UBips uint64
 
 const OneInBips Bips = 10000
+const OneInUBips UBips = 10000
 
 func NaturalToBips(natural int64) Bips {
 	return Bips(SaturatingMul(natural, int64(OneInBips)))
@@ -34,7 +36,15 @@ func UintMulByBips(value uint64, bips Bips) uint64 {
 }
 
 func SaturatingCastToBips(value uint64) Bips {
-	return Bips(SaturatingCast(value))
+	return Bips(SaturatingCast[int64](value))
+}
+
+func (bips UBips) Uint64() uint64 {
+	return uint64(bips)
+}
+
+func (bips Bips) Uint64() uint64 {
+	return uint64(bips)
 }
 
 // BigDivToBips returns dividend/divisor as bips, saturating if out of bounds
